Create app and organization separately
[#53574603]
